matchmaker combined

+++ libCacheSim/cache/eviction/GLCache/segment.c
@@ -110,7 +110,7 @@ segment_t *allocate_new_seg(cache_t *cache, int bucket_id) {
 
   new_seg->n_seg_util_skipped = 0;
   new_seg->selected_for_training = false;
-  new_seg->pred_utility = INT64_MAX;  // to avoid it being picked for eviction
+  new_seg->pred_utility = (double)INT64_MAX;  // to avoid it being picked for eviction
   new_seg->train_utility = 0;
   new_seg->magic = MAGIC;
   new_seg->seg_id = params->n_allocated_segs++;
@@ -122,8 +122,7 @@ segment_t *allocate_new_seg(cache_t *cache, int bucket_id) {
 
 /* link a segment before another segment, this is used to place the merged
  * segment in the same position as old segments */
-void link_new_seg_before_seg(GLCache_params_t *params, bucket_t *bucket,
-                             segment_t *old_seg, segment_t *new_seg) {
+void link_new_seg_before_seg(GLCache_params_t *params, bucket_t *bucket, segment_t *old_seg, segment_t *new_seg) {
   DEBUG_ASSERT(new_seg->bucket_id == bucket->bucket_id);
   DEBUG_ASSERT(old_seg->next_seg->bucket_id == bucket->bucket_id);
 
@@ -143,8 +142,7 @@ void link_new_seg_before_seg(GLCache_params_t *params, bucket_t *bucket,
 
 /* find the cutoff object score to retain objects, this is used in merging
  * multiple segments into one */
-double find_cutoff(cache_t *cache, obj_score_type_e obj_score_type,
-                   segment_t **segs, int n_segs, int n_retain) {
+double find_cutoff(cache_t *cache, obj_score_type_e obj_score_type, segment_t **segs, int n_segs, int n_retain) {
   GLCache_params_t *params = cache->eviction_params;
 
   segment_t *seg;
@@ -157,17 +155,13 @@ double find_cutoff(cache_t *cache, obj_score_type_e obj_score_type,
       double r = 1.0;
       // if (params->type != LOGCACHE_TWO_ORACLE || params->type !=
       // LOGCACHE_ITEM_ORACLE)
-      r = 1 + ((double)(next_rand() % RAND_MAX) / ((double)(RAND_MAX)) - 0.5) *
-                  0.001;
+      r = 1 + ((double)(next_rand() % RAND_MAX) / ((double)(RAND_MAX)) - 0.5) * 0.001;
 
-      params->obj_sel.score_array[pos] =
-          cal_obj_score(params, obj_score_type, &seg->objs[j]) * r;
-      params->obj_sel.score_array_offset[pos] =
-          params->obj_sel.score_array[pos];
+      params->obj_sel.score_array[pos] = cal_obj_score(params, obj_score_type, &seg->objs[j]) * r;
+      params->obj_sel.score_array_offset[pos] = params->obj_sel.score_array[pos];
       pos++;
 #else
-      params->obj_sel.score_array[pos++] =
-          cal_obj_score(params, obj_score_type, &seg->objs[j]);
+      params->obj_sel.score_array[pos++] = cal_obj_score(params, obj_score_type, &seg->objs[j]);
 #endif
     }
   }
@@ -192,14 +186,11 @@ double cal_seg_utility(cache_t *cache, segment_t *seg, bool oracle_obj_sel) {
   dd_pair_t *obj_score_online_oracle = params->obj_sel.dd_pair_array;
   for (int j = 0; j < seg->n_obj; j++) {
     if (oracle_obj_sel) {
-      obj_score_online_oracle[j].x =
-          cal_obj_score(params, OBJ_SCORE_ORACLE, &seg->objs[j]);
+      obj_score_online_oracle[j].x = cal_obj_score(params, OBJ_SCORE_ORACLE, &seg->objs[j]);
       obj_score_online_oracle[j].y = obj_score_online_oracle[j].x;
     } else {
-      obj_score_online_oracle[j].x =
-          cal_obj_score(params, params->obj_score_type, &seg->objs[j]);
-      obj_score_online_oracle[j].y =
-          cal_obj_score(params, OBJ_SCORE_ORACLE, &seg->objs[j]);
+      obj_score_online_oracle[j].x = cal_obj_score(params, params->obj_score_type, &seg->objs[j]);
+      obj_score_online_oracle[j].y = cal_obj_score(params, OBJ_SCORE_ORACLE, &seg->objs[j]);
     }
   }
 
@@ -207,8 +198,7 @@ double cal_seg_utility(cache_t *cache, segment_t *seg, bool oracle_obj_sel) {
   int n_retained_obj = 0;
 #ifdef EVICTION_CONSIDER_RETAIN
   n_retained_obj = params->n_retain_per_seg;
-  qsort(obj_score_online_oracle, seg->n_obj, sizeof(dd_pair_t),
-        cmp_double_double_pair);
+  qsort(obj_score_online_oracle, seg->n_obj, sizeof(dd_pair_t), cmp_double_double_pair);
   // DEBUG_ASSERT(obj_score_online_oracle[0].x <=
   // obj_score_online_oracle[seg->n_obj - 1].x);
 #endif
@@ -240,12 +230,10 @@ void print_seg(cache_t *cache, segment_t *seg, int log_level) {
       "mean freq %4.2lf, total hit %6d, total active %4d, "
       "%2d merges, utility %.4lf/%.4lf/%.4lf, %d obj have reuse, "
       "n_hit/active window %d %d %d %d, \n",
-      seg->seg_id, (int)params->curr_rtime - seg->create_rtime,
-      (double)seg->n_byte / seg->n_obj, seg->req_rate, seg->write_rate,
-      seg->miss_ratio, (double)seg->n_hit / seg->n_obj, seg->n_hit,
-      seg->n_active, seg->n_merge, seg->train_utility, seg->pred_utility,
-      cal_seg_utility(cache, seg, true), count_n_obj_reuse(cache, seg),
+      seg->seg_id, (int)params->curr_rtime - seg->create_rtime, (double)seg->n_byte / seg->n_obj, seg->req_rate,
+      seg->write_rate, seg->miss_ratio, (double)seg->n_hit / seg->n_obj, seg->n_hit, seg->n_active, seg->n_merge,
+      seg->train_utility, seg->pred_utility, cal_seg_utility(cache, seg, true), count_n_obj_reuse(cache, seg),
 
-      seg->feature.n_hit_per_min[0], seg->feature.n_hit_per_min[1],
-      seg->feature.n_hit_per_min[2], seg->feature.n_hit_per_min[3]);
+      seg->feature.n_hit_per_min[0], seg->feature.n_hit_per_min[1], seg->feature.n_hit_per_min[2],
+      seg->feature.n_hit_per_min[3]);
 }
